requirements.txt
azure-ai-textanalytics==5.3.0
azure-core==1.30.2
python-dotenv==1.0.1
pandas==2.2.3
tqdm==4.66.5
scikit-learn==1.5.2
